refactor(cli): split workflow.py (1050 LOC) into workflow/ subpackage (#193)
`cli/workflow.py` packed the orchestrator (`_execute_workflow`), 12
helpers across 5 unrelated concerns (loading, sweep, mounts, guards,
notifications), and the standalone Typer app into one ~1050-line module.
Splits along those concern seams:

  src/srunx/cli/workflow/
  ├── __init__.py        (no re-exports — same #194 ethos)
  ├── orchestrator.py    _execute_workflow + run_command + app + main
  ├── loading.py         _load_yaml_sweep + _resolve_endpoint_id
  ├── sweep.py           expand/locked_mounts/preview/validate_all/run_sweep
  ├── mounts.py          _resolve_sync_flag + _in_place_context + _hold_workflow_mounts
  ├── guards.py          _enforce_shell_script_roots_cli + _warn_missing_mounts
  └── notifications.py   _build_workflow_callbacks

`__init__.py` deliberately does NOT re-export. Re-exporting helpers
(``WorkflowRunner``, ``SweepOrchestrator``, ``_execute_workflow``)
shadows the submodule attributes and routes ``@patch`` calls to the wrong
site, breaking tests in subtle ways — the same conclusion #194 reached
for ``cli/commands/jobs/__init__.py``.

Updated import sites:
  - `cli/main.py:163`: `from srunx.cli.workflow.orchestrator import _execute_workflow`
  - 3 docstring cross-refs in `runtime/rendering.py`, `slurm/ssh.py`,
    `web/services/_submission_common.py` updated to the new
    `cli.workflow.mounts._hold_workflow_mounts` location.

Tests reorganized to mirror src 1:1 (CLAUDE.md test layout rules):

  tests/cli/workflow/
  ├── test_orchestrator.py            (was test_workflow.py)
  ├── test_orchestrator_sweep.py      (was test_workflow_sweep.py)
  ├── test_orchestrator_in_place.py   (was test_workflow_in_place.py)
  ├── test_orchestrator_sweep_in_place.py (was test_workflow_sweep_in_place.py)
  └── test_orchestrator_transport.py  (was test_workflow_transport.py)

Git rename detection preserved history for all 5 test files + the source
file (`workflow.py` → `workflow/orchestrator.py`).

`@patch` paths updated to target the actual lookup sites:
  - `srunx.cli.workflow.WorkflowRunner` → `srunx.cli.workflow.orchestrator.WorkflowRunner`
  - `srunx.cli.workflow.SweepOrchestrator` → `srunx.cli.workflow.sweep.SweepOrchestrator`
  - `srunx.cli.workflow.NotificationWatchCallback` → `srunx.cli.workflow.notifications.NotificationWatchCallback`
  - `srunx.cli.workflow._execute_workflow` → `srunx.cli.workflow.orchestrator._execute_workflow`

Side cleanups (deferral patterns):
  - `import yaml  # type: ignore` → `import yaml` (`types-pyyaml` is in dev deps)
  - `_load_yaml_sweep`'s broad `except Exception` narrowed to `ValueError`
    (Pydantic v2 ValidationError extends ValueError)

Verified: 2042 tests pass, ruff + mypy clean (183 source files).

Refs #193 (2 of 4 module splits remaining: ssh/core/client.py, slurm/ssh.py).
